Split viewer protocol roundtrip test
